Handle refscan pfiles with '.7' ext

+++ archive_to_folder_reaper.py
@@ -149,9 +149,9 @@ def extract_pfiles(files):
         for f in pfile_arcs:
             utd = untar(f, os.path.dirname(f))
             [_files, _dirs, _, _, _] = get_paths(utd)
-            [shutil.move(ff, os.path.dirname(utd)) for ff in _files if ff.endswith('.dat')]
+            [shutil.move(ff, os.path.dirname(utd)) for ff in _files if ff.endswith('.dat') or ff.endswith('_refscan.7')]
             for p in _files:
-                if p.endswith('.7'):
+                if p.endswith('.7') and not p.endswith('_refscan.7'):
                     gzfile = create_gzip(p, p + '.gz')
                     shutil.move(gzfile, os.path.dirname(utd))
                     shutil.rmtree(utd)
